Make HDFS_NO_IMPERSONATION tests to be not PROFILE_SPECIFIC_TESTS
There is nothing environment specific in required in the environment
to run HDFS_NO_IMPERSONATION, because hdfs impersonation is disabled by
default.

Thanks to this there is no point to run singlenode environment
because it was running only HDFS_NO_IMPERSONATION tests which is now
also executed in multinode environment.